Display errors in web dashboard, clean up and reorganize CSS/Classes
* Refactor and cleanup classes: move Bugzilla functions in their own class, remove access to global constants in Utils
* Fix warnings for son/sat (missing Bugzilla references), log error when a locale is missing from mapping array
* Display different table header for lang/raw files
* Improve CSS at smaller resolutions, clean up over qualified rules, fix background disappearing when scrolling horizontally on larger pages
* Clean up phpDOC syntax in classes
